Change the properties of the summaries
There was something with the "on demand" column that didn't work well for me, instead I'm trying "running", which is the total number of running instances, on demand, spot, EMR, etc. The spot and EMR columns are still included, but the unreserved column has changed name to "reservable" – the idea is that this better explains that it's the number of instances that could be reserved, that is the number of running minus spot, EMR, and reserved.